Add large file support
Swift requires that files over 5GB be split into multiple segments, joined by
a manifest file, allowing the object to be downloaded all-at-once.

This patch teaches `upcs` how to upload using this process as wells as adding
a new option `-z` which allows the user to override the default 5GB segment
size.

Also included is a test-suite to verify the correctness of files uploaded and
downloaded using both the small and large-object functions.